Change LazyListState extensions to return State<Boolean> instead of just Boolean
This came after a discussion on slack (https://kotlinlang.slack.com/archives/CJLTWPH7S/p1629839151061400) and understanding how receiving them as state reduces the recompositions that happen. When read as Boolean, the entire AnimalsScreen composable recomposes on any change. Reading them as State<T> limits the recomposition on the scope that actually reads the state, here the Box (but because Box inline, the entire scaffold @Composable content).